@@ -3,26 +3,28 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "babel-eslint": "^8.2.3",
+    "babel-eslint": "^8.2.6",
     "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-prettier": "^2.6.0",
-    "eslint-plugin-promise": "^3.7.0",
-    "firebase": "^4.13.1",
-    "firebase-admin": "^5.12.0",
-    "firebase-functions": "^1.0.1",
-    "material-ui": "^0.20.0",
-    "moment": "^2.22.1",
-    "moment-range": "^4.0.0",
-    "node-sass": "^4.8.3",
-    "node-sass-chokidar": "^1.2.2",
-    "npm-run-all": "^4.1.2",
-    "prettier": "^1.12.1",
+    "eslint-plugin-prettier": "^2.6.2",
+    "eslint-plugin-promise": "^3.8.0",
+    "firebase": "^5.3.0",
+    "firebase-admin": "^5.13.1",
+    "firebase-functions": "^2.0.2",
+    "material-ui": "^0.20.1",
+    "moment": "^2.22.2",
+    "moment-range": "^4.0.1",
+    "node-sass": "^4.9.2",
+    "node-sass-chokidar": "^1.3.3",
+    "npm-run-all": "^4.1.3",
+    "prettier": "^1.14.0",
     "query-string": "5",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
+    "react-sortable-hoc": "^0.8.3",
+    "styled-components": "^3.3.3",
     "typeface-roboto": "0.0.54"
   },
   "scripts": {
@@ -35,7 +37,7 @@
     "eject": "react-scripts eject"
   },
   "devDependencies": {
-    "eslint": "^4.19.1",
-    "eslint-plugin-react": "^7.7.0"
+    "eslint": "^5.2.0",
+    "eslint-plugin-react": "^7.10.0"
   }
 }
