@@ -1,19 +1,19 @@
-import React, {Component} from 'react'
+import React, {Component} from "react"
 
-import {PlaceholderText} from '../../shared'
+import {PlaceholderText} from "../../shared"
 
 
 export default class Welcome extends Component {
 
-  componentWillReceiveProps({appBarRightAction}) {
-    appBarRightAction === "welcome" && window.open("mailto:info@balazsorban.com?&subject=Hiba jelentése", '_blank')
-  }
+	UNSAFE_componentWillReceiveProps({appBarRightAction}) {
+		appBarRightAction === "welcome" && window.open("mailto:info@balazsorban.com?&subject=Hiba jelentése", "_blank")
+	}
 
-  render() {
-    return (
-      <PlaceholderText>
+	render() {
+		return (
+			<PlaceholderText>
         Admin kezelőfelület
-      </PlaceholderText>
-    )
-  }
+			</PlaceholderText>
+		)
+	}
 }
