@@ -1,121 +1,115 @@
-const functions = require('firebase-functions')
-const admin = require('firebase-admin')
-admin.initializeApp()
+'use strict';
+
+const functions = require('firebase-functions');
+const admin = require('firebase-admin');
+admin.initializeApp();
 
 // ---------------------------------------------------------------------------------------------------
-// Imports from lib
-const moment = require("./lib/moment")
-const overlaps = require('./lib/overlaps')
-const email = require("./lib/email")
-const prices = require("./lib/prices")
+// Imports from lib 📘
+const moment = require("./lib/moment");
+const overlaps = require('./lib/overlaps');
+const email = require("./lib/email");
+const prices = require("./lib/prices");
+const pictures = require("./lib/pictures");
 
 // ---------------------------------------------------------------------------------------------------
-// Database refs
-const roomsRef = admin.database().ref("rooms")
-const reservationDatesRef = admin.database().ref("reservationDates")
+// Database refs 🏦
+const reservationDatesRef = admin.database().ref("reservationDates");
 
 // ---------------------------------------------------------------------------------------------------
-// Firestore refs
-const reservationRef = functions.firestore
-                          .document("reservations/{reservationId}")
-
-
+// Firestore refs 🏦
+const reservationRef = functions.firestore.document("reservations/{reservationId}");
 
 // ---------------------------------------------------------------------------------------------------
-// Helper functions
+// Helper functions ♿
 
-const updateReservationDates = (from, to, roomId, value=null) => {
-  reservationPeriod = moment.range(from, to)
-  for (let day of reservationPeriod.by('day')) {
-    reservationDatesRef
-      .child(`${day.format("YYYY/MM/DD")}/r${roomId}`)
-      .set(value)
+const updateReservationDates = (from, to, roomId, reservationId, shouldDelete = false) => {
+  console.log("Updating reservation dates...");
+  for (let day of moment.range(from, moment(to).add(1, "day")).by('day')) {
+    let newReservation = null;
+    if (!shouldDelete) {
+      newReservation = {
+        from: day.isSame(from, "day"),
+        to: day.isSame(to, "day")
+      };
+    }
+    reservationDatesRef.child(`${day.clone().format("YYYY/MM/DD")}/r${roomId}/${reservationId}`).set(newReservation);
   }
-}
+};
 
 const isEquivalent = (a, b) => {
-  const keys = Object.keys(a)
+  const keys = Object.keys(a);
   for (let i = 0; i < keys.length; i++) {
     const key = keys[i];
     if (a[key] !== b[key]) {
-      return false
+      return false;
     }
   }
-  return true
-}
-
+  return true;
+};
 
 // ---------------------------------------------------------------------------------------------------
-// Reservation handling
+// Reservation handling 🔖
+exports.reservationCreated = reservationRef.onCreate((snap, { params: { reservationId } }) => {
+  const reservation = snap.data();
+  const { from, to, roomId } = reservation;
 
-exports.reservationCreated = reservationRef.onCreate((snap, {params: {reservationId}}) => {
-  console.log("New reservation");
-  return email.reservationCreated(reservationId, snap.data())
-})
+  updateReservationDates(from, to, roomId, snap.id);
 
+  reservation.reservationId = reservationId;
 
-exports.reservationDeleted = reservationRef.onDelete((change, context) => {
-  const {email, from, to, name, roomId} = change.data()
-  updateReservationDates(from, to, roomId)
-  return email.reservationRejected(email, name, true)
-})
+  return reservation.email !== "email@email.hu" ? email.sendReservationEmails(reservation, "created") : null;
+});
 
+exports.messageIncoming = functions.database.ref("messages/{messageId}").onCreate(email.sendMessageEmails);
 
-exports.reservationChanged = reservationRef.onUpdate(({before, after}, {params: {reservationId}}) => {
-  before = before.data()
-  after = after.data()
+exports.reservationChanged = reservationRef.onUpdate(({ before, after }, { params: { reservationId } }) => {
+  before = before.data();
+  const reservation = after.data();
 
   // 🔥 Remove old dates
-  updateReservationDates(before.from, before.to, before.roomId)
+  updateReservationDates(before.from, before.to, before.roomId, reservationId, true);
   // ✨ Add new dates
-  updateReservationDates(after.from, after.to, after.roomId, after.handled ? reservationId : null)
-  
+  updateReservationDates(reservation.from, reservation.to, reservation.roomId, reservationId);
+
+  if (reservation.email === "email@email.hu") {
+    return null;
+  }
+
+  const type =
   // Reservation accepted 🎉
-  if (!before.handled && after.handled) {
-    return email.reservationAccepted(reservationId, after)
-  }
-
-  // Reservation rejected ❌
-  if (!after.handled) {
-    return email.reservationRejected(after.email, after.name)
-  }
-
+  !before.handled && reservation.handled ? "accepted" :
   // Reservation changed 🔔
-  if (!isEquivalent(before, after)) {
-    return email.reservationChanged(before, after)
-  }
-  return null
-})
+  before.handled === reservation.handled && !isEquivalent(before, reservation) ? "changed" : null;
 
-    
-// Return overlaps in a month
-exports.overlaps = functions.https
-  .onRequest((req, res) => {
-    res.header('Access-Control-Allow-Origin', '*')
-    return overlaps.overlaps(req, res)
-  })
+  reservation.reservationId = reservationId;
+  return type ? email.sendReservationEmails(reservation, type) : null;
+});
 
+exports.reservationDeleted = reservationRef.onDelete((change, { params: { reservationId } }) => {
+  const reservation = change.data();
+  const { from, to, roomId } = reservation;
+
+  updateReservationDates(from, to, roomId, reservationId, true);
+
+  reservation.reservationId = reservationId;
+
+  return reservation.email !== "email@email.hu" ? email.sendReservationEmails(reservation, "deleted") : null;
+});
+
+// Return overlaps in a month 📅
+exports.overlaps = functions.https.onRequest(overlaps.overlaps);
 
 // ---------------------------------------------------------------------------------------------------
-// Room handling
-exports.populatePrices = functions.database
-  .ref("rooms/{roomId}/prices/metadata")
-  .onUpdate((change, context) => {
-    const {roomId} = context.params
-    const {maxPeople} = change.after.val()
-    const priceTableRef = roomsRef.child(`${roomId}/prices/table`)
-    const priceTypes = ["breakfast", "halfBoard"]
-    const promises = priceTypes.map(priceType => 
-        priceTableRef
-          .child(priceType)
-          .once("value", snap => snap.val())
-      )
-    return Promise
-      .all(promises)
-      .then(values => {
-        return priceTableRef.set({
-          "breakfast": prices.generatePrices(values[0].val(), maxPeople),
-          "halfBoard": prices.generatePrices(values[1].val(), maxPeople)
-        })
-      })
-  })
+// Room handling 🏘
+exports.populatePrices = functions.database.ref("rooms/{roomId}/prices/metadata").onUpdate(prices.populatePrices);
+
+// ---------------------------------------------------------------------------------------------------
+// Picture handling 🍱
+exports.generateThumbnail = functions.storage.object().onFinalize(pictures.generateThumbnail);
+
+exports.deleteRoomPicture = functions.database.ref("rooms/{roomId}/pictures/{pictureId}").onDelete((snap, { params: { roomId } }) => pictures.deletePicture(snap, `rooms/${roomId}`));
+
+exports.deleteFoodPicture = functions.database.ref("foods/pictures/{pictureId}").onDelete(snap => pictures.deletePicture(snap, "foods"));
+
+exports.deleteEventsPicture = functions.database.ref("events/pictures/{pictureId}").onDelete(snap => pictures.deletePicture(snap, "events"));
