@@ -1,7 +1,9 @@
-// Init moment.js
-const Moment = require('moment')
-const MomentRange = require('moment-range');
-const moment = MomentRange.extendMoment(Moment)
-moment.locale("hu")
+'use strict';
 
-module.exports = moment
+// Init moment.js
+const Moment = require('moment');
+const MomentRange = require('moment-range');
+const moment = MomentRange.extendMoment(Moment);
+moment.locale("hu");
+
+module.exports = moment;
