@@ -1,5 +1,6 @@
 import React, {Component} from "react"
 import moment from "moment"
+import QueryString from "query-string"
 import {Link, withRouter} from "react-router-dom"
 
 import {List, ListItem} from "material-ui/List"
@@ -7,8 +8,8 @@ import {List, ListItem} from "material-ui/List"
 import Subheader from "material-ui/Subheader"
 import RaisedButton from "material-ui/RaisedButton"
 
-import Edit from "material-ui/svg-icons/image/edit"
 import Done from "material-ui/svg-icons/action/done"
+import Edit from "material-ui/svg-icons/image/edit"
 import Delete from "material-ui/svg-icons/action/delete"
 import From from "material-ui/svg-icons/action/flight-land"
 import To from "material-ui/svg-icons/action/flight-takeoff"
@@ -17,11 +18,13 @@ import Tel from "material-ui/svg-icons/communication/call"
 import Email from "material-ui/svg-icons/communication/email"
 import Message from "material-ui/svg-icons/communication/message"
 import Home from "material-ui/svg-icons/action/home"
-import People from "material-ui/svg-icons/social/people"
+import Adult from "material-ui/svg-icons/social/people"
+import Child from "material-ui/svg-icons/places/child-care"
+import Money from "material-ui/svg-icons/editor/attach-money"
 
 import {AUTH,  RESERVATIONS_FS, TIMESTAMP, ADMINS } from "../../../utils/firebase"
 import {Post, ModalDialog} from "../../shared"
-import { RESERVATIONS, EDIT } from "../../../utils/routes"
+import { RESERVATIONS, EDIT, HANDLED } from "../../../utils/routes"
 
 class Reservation extends Component {
   
@@ -75,19 +78,21 @@ class Reservation extends Component {
   		isDeleting,
   		reservation: {
   			id,
-  			name, email, tel, message, adults, children, 
+				name, email, tel, message, adults, children,
+				price=0,
   			roomId, from, to, handled,
 				timestamp,
 				lastHandledBy
   		}
   	} = this.state
-    
 		
-    
+		const expanded = this.props.location.search.includes(id)
+		const handledStatus = QueryString.parse(this.props.location.search).kezelt || "nem"
   	return (
   		<ListItem disabled style={{margin: "0 auto", padding: 0}}>
   			<Post
-  				title={`${name} (${moment(to).diff(moment(from), "days")+1} nap)`}
+					{...{expanded}}
+  				title={`${name} (${moment(to && (to.seconds*1000 || to)).diff(moment(from && (from.seconds*1000 || from)), "days")+1} nap)`}
   				subtitle={`Szoba ${roomId}`}
   				rightText={<p style={{margin: "-2.5em 2.5em 0 0", textAlign: "right", color: "#aaa", fontSize: ".8em", fontStyle: "italic"}}>{!lastHandledBy ? "Foglalás dátuma" : `módosítva (${lastHandledBy})`}<br/> {moment(timestamp ? moment.unix(timestamp.seconds) : undefined).format("YYYY. MMMM DD. HH:mm:ss")}</p>}
   			> 
@@ -117,23 +122,34 @@ class Reservation extends Component {
   						secondaryText="szoba"
   					/>
   					<ListItem disabled
-  						leftIcon={<People/>}
+  						leftIcon={<Adult/>}
   						primaryText={adults} 
   						secondaryText="felnőtt"
   					/>
   					<ListItem disabled
-  						leftIcon={<People/>}
+  						leftIcon={<Child/>}
   						primaryText={(children && children.length) || "0"} 
   						secondaryText="gyerek"
   					/>
+						<ListItem disabled
+  						leftIcon={<Money/>}
+  						primaryText={
+								(price)
+									.toLocaleString("hu-HU", {
+										style: "currency", currency: "HUF",
+										maximumFractionDigits: 0,
+										minimumFractionDigits: 0
+									})} 
+  						secondaryText="fizetni"
+  					/>
   					<ListItem disabled
   						leftIcon={<From/>}
-  						primaryText={moment(from).format("MMMM D. HH:mm")} 
+  						primaryText={moment(from && (from.seconds*1000 || from)).format("MMMM D. HH:mm")} 
   						secondaryText="érkezés"
   					/>
   					<ListItem disabled
   						leftIcon={<To/>}
-  						primaryText={moment(to).format("MMMM D. HH:mm")} 
+  						primaryText={moment(to && (to.seconds*1000 || to)).format("MMMM D. HH:mm")} 
   						secondaryText="távozás"
   					/>
   				</List>
@@ -148,7 +164,7 @@ class Reservation extends Component {
                 	onClick={() => this.handleReservation(true)}
                 />
   						}
-  						<Link to={`${RESERVATIONS}/${id}/${EDIT}`}>
+  						<Link to={`${RESERVATIONS}/${id}/${EDIT}?kezelt=${handledStatus}`}>
   							<RaisedButton
   								style={{margin: "0 12px"}}
   								label={window.innerWidth >= 640 && "Szerkeszt"}
