@@ -1,5 +1,7 @@
-import firebase from "firebase"
+import firebase from "firebase/app"
+import "firebase/database"
 import "firebase/firestore"
+import "firebase/auth"
 
 firebase.initializeApp({
 	apiKey: "AIzaSyB4-Y2_RCdrOouJJxUJkBBXGyj4hNdjDs0",
@@ -13,24 +15,20 @@ const firestore = firebase.firestore()
 const settings = {timestampsInSnapshots: true}
 firestore.settings(settings)
 
-const DB = firebase.database()
-const FS = firebase.firestore()
-const AUTH = firebase.auth()
+export const DB = firebase.database()
+export const FS = firebase.firestore()
+export const AUTH = firebase.auth()
 
 
 
-const RESERVATIONS_FS = FS.collection("reservations")
-const FEEDBACKS_DB = DB.ref("feedbacks")
-const ADMINS = DB.ref("admins")
-const ROOMS_DB = DB.ref("rooms")
-const ROOM_SERVICES_DB = DB.ref("roomServices")
-const SERVER_MESSAGE_DB = DB.ref("serverMessage")
+export const RESERVATIONS_FS = FS.collection("reservations")
+export const FEEDBACKS_DB = DB.ref("feedbacks")
+export const ADMINS = DB.ref("admins")
+export const ROOMS_DB = DB.ref("rooms")
+export const FOODS_DB = DB.ref("foods")
+export const ROOM_SERVICES_DB = DB.ref("roomServices")
+export const SERVER_MESSAGE_DB = DB.ref("serverMessage")
+export const RESERVATION_DATES_DB = DB.ref("reservationDates")
+export const PARAGRAPHS_DB = DB.ref("paragraphs")
 
-const TIMESTAMP = firebase.firestore.FieldValue.serverTimestamp()
-
-
-export {DB, FS, AUTH, 
-	RESERVATIONS_FS, FEEDBACKS_DB,
-	ROOM_SERVICES_DB, ROOMS_DB, SERVER_MESSAGE_DB,
-	ADMINS, TIMESTAMP
-}
+export const TIMESTAMP = firebase.firestore.FieldValue.serverTimestamp()
