@@ -1,6 +1,6 @@
-import moment from 'moment'
-import 'moment/locale/hu'
-import {extendMoment} from 'moment-range'
+import moment from "moment"
+import "moment/locale/hu"
+import {extendMoment} from "moment-range"
 extendMoment(moment)
 
 // This file is related to moment.js
