@@ -1,13 +1,16 @@
 export const WELCOME = "/kezdolap"
 export const SPECIAL_OFFER = "/ajanlatok"
 export const ROOMS = "/szobak"
+export const INTRO = "/bemutatkozas"
+export const CERTIFICATES = "/tanusitvanyok"
 export const EVENTS = "/rendezvenyek"
 export const FOODS = "/etelek"
 export const SETTINGS = "/beallitasok"
 export const CALENDAR = "/naptar"
 export const RESERVATIONS = "/foglalasok"
 export const FEEDBACKS = "/visszajelzesek"
-export const STATS = "/statisztika"
+export const STATS = "/statisztikak"
 
 
 export const EDIT = "szerkeszt"
+export const HANDLED = "kezelt"
